Extracted screen file from current screen and erased duplicate theme package.
