+++ app/src/main/java/com/alexcfa/yourweather/ui/screens/Screen.kt
@@ -1,0 +1,19 @@
+package com.alexcfa.yourweather.ui.screens
+
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Surface
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import com.alexcfa.yourweather.ui.theme.YourWeatherTheme
+
+@Composable
+fun Screen(content: @Composable () -> Unit) {
+    YourWeatherTheme {
+        Surface(
+            modifier = Modifier.fillMaxSize(),
+            color = MaterialTheme.colorScheme.background,
+            content = content
+        )
+    }
+}
